+++ builder_float.go
@@ -5,8 +5,8 @@ import (
 	"math"
 	"strconv"
 
+	"github.com/dogmatiq/ferrite/internal/maybe"
 	"github.com/dogmatiq/ferrite/internal/reflectx"
-	"github.com/dogmatiq/ferrite/maybe"
 	"github.com/dogmatiq/ferrite/variable"
 	"golang.org/x/exp/constraints"
 )