+++ internal/maybe/doc.go
@@ -2,8 +2,4 @@
 //
 // We use the "maybe" terminology so as to avoid confusion with the "optional"
 // and "required" types used to represent environment variable values.
-//
-// WARNING: This package is exposed to allow for experimentation but it is not
-// yet considered part of Ferrite's public API. It may change at any time and
-// without warning.
 package maybe